fin mongo

+++ IT_Academy/Java BackEnd & Spring/Sprint_5/S5T01N02/MONGO_DB/src/main/java/com/itacademy/barcelonactiva/HornosJordi/S5T01N02/model/dto/PlayerDTO.java
@@ -11,9 +11,9 @@ public class PlayerDTO {
 
 
     private String username;
-    private Float winRate;
+    private Double winRate;
 
-    public PlayerDTO(String username, Float winRate){
+    public PlayerDTO(String username, Double winRate){
         this.username = Objects.requireNonNull(username, "UNKNOWN");
         this.winRate = winRate;
     }
